move function CheckTxnAuth from package txn to apply
