+++ server/etcdserver/apply/auth_test.go
@@ -33,6 +33,7 @@ import (
 	"go.etcd.io/etcd/server/v3/etcdserver/api/v3alarm"
 	"go.etcd.io/etcd/server/v3/etcdserver/cindex"
 	"go.etcd.io/etcd/server/v3/lease"
+	"go.etcd.io/etcd/server/v3/storage/backend"
 	betesting "go.etcd.io/etcd/server/v3/storage/backend/testing"
 	"go.etcd.io/etcd/server/v3/storage/mvcc"
 	"go.etcd.io/etcd/server/v3/storage/schema"
@@ -820,3 +821,367 @@ func TestCheckLeasePutsKeys(t *testing.T) {
 	aa.authInfo = auth.AuthInfo{Username: "bob", Revision: aa.as.Revision()}
 	assert.NoErrorf(t, aa.checkLeasePutsKeys(l), "bob should be able to access key 'a'")
 }
+
+func TestCheckTxnAuth(t *testing.T) {
+	be, _ := betesting.NewDefaultTmpBackend(t)
+	defer betesting.Close(t, be)
+	as := setupAuth(t, be)
+
+	tests := []struct {
+		name       string
+		txnRequest *pb.TxnRequest
+		err        error
+	}{
+		{
+			name: "Out of range compare is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Compare: []*pb.Compare{outOfRangeCompare},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "In range compare is authorized",
+			txnRequest: &pb.TxnRequest{
+				Compare: []*pb.Compare{inRangeCompare},
+			},
+			err: nil,
+		},
+		{
+			name: "Nil request range is always authorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{nilRequestRange},
+			},
+			err: nil,
+		},
+		{
+			name: "Range request in range is authorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{inRangeRequestRange},
+				Failure: []*pb.RequestOp{inRangeRequestRange},
+			},
+			err: nil,
+		},
+		{
+			name: "Range request out of range success case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{outOfRangeRequestRange},
+				Failure: []*pb.RequestOp{inRangeRequestRange},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Range request out of range failure case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{inRangeRequestRange},
+				Failure: []*pb.RequestOp{outOfRangeRequestRange},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Nil Put request is always authorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{nilRequestPut},
+			},
+			err: nil,
+		},
+		{
+			name: "Put request in range in authorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{inRangeRequestPut},
+				Failure: []*pb.RequestOp{inRangeRequestPut},
+			},
+			err: nil,
+		},
+		{
+			name: "Put request out of range success case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{outOfRangeRequestPut},
+				Failure: []*pb.RequestOp{inRangeRequestPut},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Put request out of range failure case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{inRangeRequestPut},
+				Failure: []*pb.RequestOp{outOfRangeRequestPut},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Nil delete request is authorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{nilRequestDeleteRange},
+			},
+			err: nil,
+		},
+		{
+			name: "Delete range request in range is authorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{inRangeRequestDeleteRange},
+				Failure: []*pb.RequestOp{inRangeRequestDeleteRange},
+			},
+			err: nil,
+		},
+		{
+			name: "Delete range request out of range success case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{outOfRangeRequestDeleteRange},
+				Failure: []*pb.RequestOp{inRangeRequestDeleteRange},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Delete range request out of range failure case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{inRangeRequestDeleteRange},
+				Failure: []*pb.RequestOp{outOfRangeRequestDeleteRange},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Delete range request out of range and PrevKv false success case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{outOfRangeRequestDeleteRangeKvFalse},
+				Failure: []*pb.RequestOp{inRangeRequestDeleteRange},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Delete range request out of range and PrevKv false failure case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{inRangeRequestDeleteRange},
+				Failure: []*pb.RequestOp{outOfRangeRequestDeleteRangeKvFalse},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Nested txn request in range is authorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{
+					{
+						Request: &pb.RequestOp_RequestTxn{
+							RequestTxn: &pb.TxnRequest{
+								Success: []*pb.RequestOp{inRangeRequestRange, inRangeRequestPut},
+								Failure: []*pb.RequestOp{inRangeRequestDeleteRange},
+							},
+						},
+					},
+				},
+			},
+			err: nil,
+		},
+		{
+			name: "Nested txn request out of range success case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{
+					{
+						Request: &pb.RequestOp_RequestTxn{
+							RequestTxn: &pb.TxnRequest{
+								Success: []*pb.RequestOp{outOfRangeRequestRange},
+							},
+						},
+					},
+				},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Nested txn request out of range failure case is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Failure: []*pb.RequestOp{
+					{
+						Request: &pb.RequestOp_RequestTxn{
+							RequestTxn: &pb.TxnRequest{
+								Failure: []*pb.RequestOp{outOfRangeRequestPut},
+							},
+						},
+					},
+				},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Nested txn request out of range delete is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{
+					{
+						Request: &pb.RequestOp_RequestTxn{
+							RequestTxn: &pb.TxnRequest{
+								Success: []*pb.RequestOp{outOfRangeRequestDeleteRange},
+							},
+						},
+					},
+				},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+		{
+			name: "Two level nested txn request out of range delete is unauthorized",
+			txnRequest: &pb.TxnRequest{
+				Success: []*pb.RequestOp{
+					{
+						Request: &pb.RequestOp_RequestTxn{
+							RequestTxn: &pb.TxnRequest{
+								Failure: []*pb.RequestOp{
+									{
+										Request: &pb.RequestOp_RequestTxn{
+											RequestTxn: &pb.TxnRequest{
+												Success: []*pb.RequestOp{outOfRangeRequestDeleteRange},
+											},
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			err: auth.ErrPermissionDenied,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := CheckTxnAuth(as, &auth.AuthInfo{Username: "foo", Revision: 8}, tt.txnRequest)
+			assert.Equal(t, tt.err, err)
+		})
+	}
+}
+
+// CheckTxnAuth test setup.
+func setupAuth(t *testing.T, be backend.Backend) auth.AuthStore {
+	lg := zaptest.NewLogger(t)
+
+	simpleTokenTTLDefault := 300 * time.Second
+	tokenTypeSimple := "simple"
+	dummyIndexWaiter1 := func(index uint64) <-chan struct{} {
+		ch := make(chan struct{}, 1)
+		go func() {
+			ch <- struct{}{}
+		}()
+		return ch
+	}
+
+	tp, _ := auth.NewTokenProvider(zaptest.NewLogger(t), tokenTypeSimple, dummyIndexWaiter1, simpleTokenTTLDefault)
+
+	as := auth.NewAuthStore(lg, schema.NewAuthBackend(lg, be), tp, 4)
+
+	// create "root" user and "foo" user with limited range
+	_, err := as.RoleAdd(&pb.AuthRoleAddRequest{Name: "root"})
+	require.NoError(t, err)
+
+	_, err = as.RoleAdd(&pb.AuthRoleAddRequest{Name: "rw"})
+	require.NoError(t, err)
+
+	_, err = as.RoleGrantPermission(&pb.AuthRoleGrantPermissionRequest{
+		Name: "rw",
+		Perm: &authpb.Permission{
+			PermType: authpb.Permission_READWRITE,
+			Key:      []byte("foo"),
+			RangeEnd: []byte("zoo"),
+		},
+	})
+	require.NoError(t, err)
+
+	_, err = as.UserAdd(&pb.AuthUserAddRequest{Name: "root", Password: "foo"})
+	require.NoError(t, err)
+
+	_, err = as.UserAdd(&pb.AuthUserAddRequest{Name: "foo", Password: "foo"})
+	require.NoError(t, err)
+
+	_, err = as.UserGrantRole(&pb.AuthUserGrantRoleRequest{User: "root", Role: "root"})
+	require.NoError(t, err)
+
+	_, err = as.UserGrantRole(&pb.AuthUserGrantRoleRequest{User: "foo", Role: "rw"})
+	require.NoError(t, err)
+
+	err = as.AuthEnable()
+	require.NoError(t, err)
+
+	return as
+}
+
+// CheckTxnAuth variables setup.
+var (
+	inRangeCompare = &pb.Compare{
+		Key:      []byte("foo"),
+		RangeEnd: []byte("zoo"),
+	}
+	outOfRangeCompare = &pb.Compare{
+		Key:      []byte("boo"),
+		RangeEnd: []byte("zoo"),
+	}
+	nilRequestPut = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestPut{
+			RequestPut: nil,
+		},
+	}
+	inRangeRequestPut = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestPut{
+			RequestPut: &pb.PutRequest{
+				Key: []byte("foo"),
+			},
+		},
+	}
+	outOfRangeRequestPut = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestPut{
+			RequestPut: &pb.PutRequest{
+				Key: []byte("boo"),
+			},
+		},
+	}
+	nilRequestRange = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestRange{
+			RequestRange: nil,
+		},
+	}
+	inRangeRequestRange = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestRange{
+			RequestRange: &pb.RangeRequest{
+				Key:      []byte("foo"),
+				RangeEnd: []byte("zoo"),
+			},
+		},
+	}
+	outOfRangeRequestRange = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestRange{
+			RequestRange: &pb.RangeRequest{
+				Key:      []byte("boo"),
+				RangeEnd: []byte("zoo"),
+			},
+		},
+	}
+	nilRequestDeleteRange = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestDeleteRange{
+			RequestDeleteRange: nil,
+		},
+	}
+	inRangeRequestDeleteRange = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestDeleteRange{
+			RequestDeleteRange: &pb.DeleteRangeRequest{
+				Key:      []byte("foo"),
+				RangeEnd: []byte("zoo"),
+				PrevKv:   true,
+			},
+		},
+	}
+	outOfRangeRequestDeleteRange = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestDeleteRange{
+			RequestDeleteRange: &pb.DeleteRangeRequest{
+				Key:      []byte("boo"),
+				RangeEnd: []byte("zoo"),
+				PrevKv:   true,
+			},
+		},
+	}
+	outOfRangeRequestDeleteRangeKvFalse = &pb.RequestOp{
+		Request: &pb.RequestOp_RequestDeleteRange{
+			RequestDeleteRange: &pb.DeleteRangeRequest{
+				Key:      []byte("boo"),
+				RangeEnd: []byte("zoo"),
+				PrevKv:   false,
+			},
+		},
+	}
+)
